% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcResiduals}
\alias{calcResiduals}
\title{Calculate the residuals}
\usage{
calcResiduals(Y, X, Z, gamma, beta, alpha, intercept = FALSE, exo = NULL)
}
\arguments{
\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{gamma}{Estimate of short-run effects, or an integer specifying the number of lags to initialise it here.}

\item{beta}{Estimate of the cointegrating vector}

\item{alpha}{Estimate of the adjustment coefficients}

\item{intercept}{Should an intercept be added to the data, default is FALSE}
}
\value{
The residuals
}
\description{
Calculate the residuals
}
