% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkExo}
\alias{checkExo}
\title{Check the formatting of exogenous data}
\usage{
checkExo(exo, n_obs, exo_K = NULL)
}
\arguments{
\item{exo}{The exogenous data to check}

\item{n_obs}{The number of observations it should have}

\item{exo_K}{The number of exogenous variables exo should have}
}
\description{
Check the formatting of exogenous data
}
