% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ciPlot}
\alias{ciPlot}
\title{Create a confidence interval plot for a forecast}
\usage{
ciPlot(
  data,
  col = 1,
  ylim = c(min(data$low[, col]), max(data$high[, col])),
  ci_col = "lightgrey",
  ...
)
}
\arguments{
\item{data}{The data to plot, should be a list that contains "fcst", "low" and "high" matrices}

\item{col}{Which column of each matrix contains the data to plot}

\item{ylim}{The y axis limits for the plot}

\item{ci_col}{The confidence area color}

\item{...}{Additional arguments to pass to the plot function.}
}
\description{
Create a confidence interval plot for a forecast
}
