% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crossValidate}
\alias{crossValidate}
\title{Run cross validation to determine the optimal lambda value}
\usage{
crossValidate(cutoff, Ymatrix, Xmatrix, lambdas, glmnetthresh)
}
\description{
Run cross validation to determine the optimal lambda value
}
