% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{determineRank}
\alias{determineRank}
\title{Function to determine the cointegration rank using the rank selection criterion}
\usage{
determineRank(
  data,
  r.init = NULL,
  p,
  max.iter.lasso = 3,
  conv.lasso = 10^-2,
  max.iter.r = 5,
  beta.init,
  alpha.init,
  rho.glasso = 0.1,
  lambda.gamma = matrix(seq(from = 0.1, to = 0.001, length = 10), nrow = 1),
  lambda_beta = matrix(seq(from = 2, to = 0.001, length = 100), nrow = 1),
  glmnetthresh = 1e-04
)
}
\arguments{
\item{r.init}{initial value of cointegration rank}

\item{p}{number of lags to be included}

\item{max.iter.lasso}{maximum number of iterations PML}

\item{conv.lasso}{convergence parameter}

\item{max.iter.r}{maximu number of iterations to compute cointegration rank}

\item{beta.init}{initial value for beta}

\item{alpha.init}{initial value for alpha}

\item{rho.glasso}{tuning parameter for inverse covariance matrix}

\item{lambda.gamma}{tuning parameter for GAMMA}

\item{glmnetthresh}{tolerance parameter glmnet function}

\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{lambda.beta}{tuning parameter for BETA}
}
\value{
A List containing;
rhat: estimated cointegration rank
it.r: number of iterations
rhat_iterations: estimate of cointegration rank in each iteration
mu: value of mu
decomp: eigenvalue decomposition
}
\description{
Function to determine the cointegration rank using the rank selection criterion
}
