% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint.R
\name{fitted.sparsecoint}
\alias{fitted.sparsecoint}
\title{Calculate fitted values as single step ahead forecasts}
\usage{
\method{fitted}{sparsecoint}(model)
}
\description{
Calculate fitted values as single step ahead forecasts
}
