% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginv.R
\name{ginv}
\alias{ginv}
\title{Moore-Penrose pseudoinverse adapted from the package MASS}
\usage{
ginv(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{The matrix to (pseudo)invert}

\item{tol}{The tolerance to use}
}
\value{
The Moore-Penrose pseudoinverse of x
}
\description{
Moore-Penrose pseudoinverse adapted from the package MASS
}
