% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lagNames}
\alias{lagNames}
\title{Create names for lagged variables for summary}
\usage{
lagNames(varnames, p, intercept = FALSE)
}
\arguments{
\item{varnames}{A vector of variable names}

\item{p}{The number of lags}

\item{intercept}{If the names should contain an intercept variable}
}
\value{
A character vector containing the lagged variable names
}
\description{
Create names for lagged variables for summary
}
