% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matrixQuantiles}
\alias{matrixQuantiles}
\title{Function for calculating the quantiles and mean of many matrices in a list}
\usage{
matrixQuantiles(forecast, matlist, quantiles = c(0.025, 0.975))
}
\arguments{
\item{matlist}{The list of matrices}

\item{quantiles}{The quantiles to calculate}
}
\value{
A list containing fcst (the forecast) and low and high being the quantiles
}
\description{
Function for calculating the quantiles and mean of many matrices in a list
}
