% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{nts.alpha.procrusted}
\alias{nts.alpha.procrusted}
\title{Function to estimate alpha}
\usage{
nts.alpha.procrusted(Y, X, Z, zbeta, r, Omega, P, beta, intercept = F)
}
\arguments{
\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{zbeta}{estimate of short-run effects}

\item{r}{cointegration rank}

\item{Omega}{estimate of inverse error covariance matrix}

\item{P}{transformation matrix P derived from Omega}

\item{beta}{estimate of cointegrating vector}

\item{intercept}{F do not include intercept, T include intercept in estimation short-run effects}
}
\value{
A list containing:
alpha: estimate of adjustment coefficients
alphastar: estimate of transformed adjustment coefficients
}
\description{
Function to estimate alpha
}
