% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{nts.gamma.fixed.lassoreg}
\alias{nts.gamma.fixed.lassoreg}
\title{Function to estimate gamma, using l1 penalty, fixed value of lambda.gamma}
\usage{
nts.gamma.fixed.lassoreg(
  Y,
  X,
  Z,
  Pi,
  Omega,
  lambda.gamma = 0.001,
  p,
  intercept = F,
  glmnetthresh = 1e-04
)
}
\arguments{
\item{Y:}{Response Time Series}

\item{X:}{Time Series in Differences}

\item{Z:}{Time Series in Levels}

\item{Pi:}{estimate of cointegration space}

\item{Omega:}{estimate of inverse error covariance matrix}

\item{lambda.gamma:}{tuning paramter short-run effects}

\item{p:}{number of lagged differences}

\item{cutoff:}{cutoff value time series cross-validation approach}

\item{intercept:}{F do not include intercept, T include intercept}

\item{glmnetthresh:}{tolerance parameter glmnet function}
}
\value{
A list containing:
zbeta: estimate of short-run effects
P: transformation matrix P derived from Omega
}
\description{
Function to estimate gamma, using l1 penalty, fixed value of lambda.gamma
}
