% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{nts.omega}
\alias{nts.omega}
\title{Determine Omega, conditional on alpha,beta and gamma}
\usage{
nts.omega(Y, X, Z, gamma, beta, alpha, rho)
}
\arguments{
\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{gamma}{Estimate of the short-run effects}

\item{beta}{Estimate of the cointegrating vector}

\item{alpha}{Estimate of the adjustment coefficients}

\item{rho}{A single or multiple values for the tuning parameter rho for the glasso algorithm.}
}
\value{
The omega matrix as estimated with glasso, and the selected value for rho.
}
\description{
Determine Omega, conditional on alpha,beta and gamma
}
