% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sparsecoint_pred}
\alias{plot.sparsecoint_pred}
\title{Plot function for sparsecoint_pred}
\usage{
\method{plot}{sparsecoint_pred}(pred, variable = 1)
}
\arguments{
\item{pred}{The prediction from a sparsecoint model}
}
\description{
Plot function for sparsecoint_pred
}
