% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{rankSelectionCriterion}
\alias{rankSelectionCriterion}
\title{Sparse cointegration function used in determine_rank function for Rank Selection Criterion}
\usage{
rankSelectionCriterion(
  p,
  Y,
  X,
  Z,
  r,
  alpha.init = NULL,
  beta.init,
  max.iter = 25,
  conv = 10^-3,
  lambda.gamma = 0.001,
  lambda_beta = matrix(seq(from = 2, to = 0.001, length = 100), nrow = 1),
  rho.glasso = 0.5,
  cutoff = 0.8,
  intercept = F,
  glmnetthresh = 1e-04
)
}
\arguments{
\item{p}{number of lagged differences}

\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{r}{cointegration rank}

\item{alpha.init}{initial value for adjustment coefficients}

\item{beta.init}{initial value for cointegrating vector}

\item{max.iter}{maximum number of iterations}

\item{conv}{convergence parameter}

\item{lambda.gamma}{tuning paramter short-run effects}

\item{lambda_beta}{tuning paramter cointegrating vector}

\item{rho.glasso}{tuning parameter inverse error covariance matrix}

\item{cutoff}{cutoff value time series cross-validation approach}

\item{glmnetthresh}{tolerance parameter glmnet function}
}
\value{
A list containing:
BETAhat: estimate of cointegrating vectors
ALPHAhat: estimate of adjustment coefficients
ZBETA: estimate of short-run effects
OMEGA: estimate of inverse covariance matrix
}
\description{
Sparse cointegration function used in determine_rank function for Rank Selection Criterion
}
