% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{rearrangeYX}
\alias{rearrangeYX}
\title{Rearrange X and Y to be in the appropriate order for cross validatuon}
\usage{
rearrangeYX(Y, X, q)
}
\arguments{
\item{Y}{The dependent variable}

\item{X}{The independent lags}

\item{q}{The number of dependent variables}
}
\value{
The Y and X matrices rearranged to be in time order
}
\description{
Rearrange X and Y to be in the appropriate order for cross validatuon
}
