% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint.R
\name{residuals.sparsecoint}
\alias{residuals.sparsecoint}
\title{Calculate fitted values as single step ahead forecasts}
\usage{
\method{fitted}{sparsecoint}(x)
}
\arguments{
\item{x}{The sparsecoint model object}
}
\description{
Calculate fitted values as single step ahead forecasts
}
