% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{restrictLambda}
\alias{restrictLambda}
\title{Restrict the lambda sequence: exclude all-zero solutions}
\usage{
restrictLambda(Ymatrixsd, Xmatrix, lambda, tol)
}
\description{
Restrict the lambda sequence: exclude all-zero solutions
}
