% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{samples.predict.sparsecoint}
\alias{samples.predict.sparsecoint}
\title{Sample many predictions from a sparesecoint model to create prediction intervals}
\usage{
samples.predict.sparsecoint(x, h = 1, exo = NULL, samples = 1)
}
\arguments{
\item{h}{The number of periods to forecast}

\item{samples}{The number of samples to obtain}

\item{model}{The sparsecoint object}
}
\value{
A list containing the raw sampled forecasts
}
\description{
Sample many predictions from a sparesecoint model to create prediction intervals
}
