% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupdata.R
\name{setupData}
\alias{setupData}
\title{Set up data for the sparsecoint model by diffing and lagging it}
\usage{
setupData(data, p = 1)
}
\description{
Set up data for the sparsecoint model by diffing and lagging it
}
