% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{shiftLag}
\alias{shiftLag}
\title{Shift lagged data one step by adding a new observation}
\usage{
shiftLag(new, lagged)
}
\arguments{
\item{new}{The new data}

\item{lagged}{The lagged data (in a single row)}
}
\value{
The lagged data shifted one lag with new as the new first lag
}
\description{
Shift lagged data one step by adding a new observation
}
