\name{sparsecoint-package}
\alias{sparsecoint-package}
\alias{sparsecoint}
\docType{package}
\title{
\packageTitle{sparsecoint}
}
\description{
\packageDescription{sparsecoint}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sparsecoint}
\packageIndices{sparsecoint}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sparsecoint}

Maintainer: \packageMaintainer{sparsecoint}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
