% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint.R
\name{sparsecoint}
\alias{sparsecoint}
\title{Create and fit a sparsecoint model}
\usage{
sparsecoint(data, p = 1, exo = NULL, exo_p = p, intercept = FALSE)
}
\description{
Create and fit a sparsecoint model
}
