% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint_lasso.R
\name{SparseCointegration_Lasso}
\alias{SparseCointegration_Lasso}
\title{Main function to perform sparse cointegration}
\usage{
SparseCointegration_Lasso(
  data,
  p,
  r,
  alpha.init = NULL,
  beta.init = NULL,
  max.iter = 10,
  conv = 10^-2,
  lambda.gamma,
  rho.glasso = seq(from = 1, to = 0.1, length = 5),
  lambda_beta,
  cutoff = 0.8,
  glmnetthresh = 1e-04
)
}
\arguments{
\item{p:}{number of lagged differences}

\item{Y:}{Response Time Series}

\item{X:}{Time Series in Differences}

\item{Z:}{Time Series in Levels}

\item{r:}{cointegration rank}

\item{alpha.init:}{initial value for adjustment coefficients}

\item{beta.init:}{initial value for cointegrating vector}

\item{max.iter:}{maximum number of iterations}

\item{conv:}{convergence parameter}

\item{lambda.gamma:}{tuning paramter short-run effects}

\item{lambda_beta:}{tuning paramter cointegrating vector}

\item{rho.glasso:}{tuning parameter inverse error covariance matrix}

\item{cutoff:}{cutoff value time series cross-validation approach}

\item{glmnetthresh:}{tolerance parameter glmnet function}
}
\value{
A list containing:
beta: estimate of cointegrating vectors
alpha: estimate of adjustment coefficients
gamma: estimate of short-run effects
omega: estimate of inverse covariance matrix
}
\description{
Main function to perform sparse cointegration
}
