% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crossValidate}
\alias{crossValidate}
\title{Run cross validation to determine the optimal lambda value}
\usage{
crossValidate(q, cutoff, Y, X, lambdas, tol)
}
\description{
Run cross validation to determine the optimal lambda value
}
