% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint.R
\name{new_sparsecoint}
\alias{new_sparsecoint}
\title{Create a new model object}
\usage{
new_sparsecoint(data, p = 1, exo = NULL, exo_p = p, intercept = FALSE)
}
\arguments{
\item{data}{The data to use}

\item{p}{The lag order to use}

\item{intercept}{Should an intercept be included, default is FALSE}
}
\value{
A new model of the class sparsecoint
}
\description{
Create a new model object
}
