% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{nts.beta}
\alias{nts.beta}
\title{Function to estimate beta and omega
First step: Determine Beta conditional on Gamma, alpha and Omega using Lasso Regression
Second step: Determine Omega conditional on Gamma, alpha and beta using GLasso}
\usage{
nts.beta(
  Y,
  X,
  Z,
  gamma,
  rank,
  P,
  alpha,
  alphastar,
  lambda = NULL,
  rho_omega,
  cutoff,
  intercept = F,
  tol = 1e-04
)
}
\arguments{
\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{gamma}{estimate of short-run effects}

\item{rank}{cointegration rank}

\item{P}{transformation matrix P derived from Omega}

\item{alpha}{estimate of adjustment coefficients}

\item{alphastar}{estimate of transformed adjustment coefficients}

\item{lambda}{tuning paramter cointegrating vector}

\item{rho_omega}{tuning parameter inverse error covariance matrix}

\item{cutoff}{cutoff value time series cross-validation approach}

\item{intercept}{F do not include intercept, T include intercept in estimation short-run effects}

\item{tol}{tolerance parameter glmnet function}

\item{Omega}{estimate of inverse error covariance matrix}
}
\value{
A list containing:
BETA: estimate of cointegrating vectors
OMEGA: estimate of inverse covariance matrix
}
\description{
Function to estimate beta and omega
First step: Determine Beta conditional on Gamma, alpha and Omega using Lasso Regression
Second step: Determine Omega conditional on Gamma, alpha and beta using GLasso
}
