% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{nts.gamma.lassoreg}
\alias{nts.gamma.lassoreg}
\title{Function to estimate gamma, using l1 penalty ###}
\usage{
nts.gamma.lassoreg(
  Y,
  X,
  Z,
  alpha,
  beta,
  Omega,
  lambda = matrix(seq(from = 1, to = 0.01, length = 10), nrow = 1),
  p,
  cutoff = 0.8,
  intercept = F,
  exo = NULL,
  tol = 1e-04,
  fixed = FALSE
)
}
\arguments{
\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}

\item{alpha}{Estimate of the adjustment coefficients}

\item{beta}{Estimate of the cointegrating vector}

\item{Omega}{estimate of inverse error covariance matrix}

\item{lambda}{tuning paramter short-run effects}

\item{p}{number of lagged differences}

\item{cutoff}{cutoff value time series cross-validation approach}

\item{intercept}{F do not include intercept, T include intercept}

\item{tol}{tolerance parameter glmnet function}
}
\value{
A list containing:
gamma: estimate of short-run effects
P: transformation matrix P derived from Omega
}
\description{
Function to estimate gamma, using l1 penalty ###
}
