% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sparsecoint}
\alias{predict.sparsecoint}
\title{Predict using a model of class sparsecoint}
\usage{
\method{predict}{sparsecoint}(x, h = 1, samples = FALSE, PI = 0.05)
}
\arguments{
\item{x}{The sparsecoint model object}

\item{h}{The prediction horizon}
}
\value{
A prediction of length h
}
\description{
Predict using a model of class sparsecoint
}
