% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{restrictLambda}
\alias{restrictLambda}
\title{Restrict the lambda sequence: exclude all zero-solution}
\usage{
restrictLambda(Ymatrixsd, Xmatrix, lambda, glmnetthresh)
}
\description{
Restrict the lambda sequence: exclude all zero-solution
}
