% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{singlestep.sparsecoint}
\alias{singlestep.sparsecoint}
\title{Single step predictions on the diff scale}
\usage{
singlestep.sparsecoint(alpha, beta, gamma, level_data, diff_lag_data)
}
\description{
Single step predictions on the diff scale
}
