% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint_lasso.R
\name{SparseCointegration_Lasso}
\alias{SparseCointegration_Lasso}
\title{Main function to perform sparse cointegration}
\usage{
SparseCointegration_Lasso(
  data,
  p,
  r,
  alpha = NULL,
  beta = NULL,
  max.iter = 10,
  conv = 10^-2,
  lambda_gamma,
  rho_omega,
  lambda_beta,
  cutoff = 0.8,
  tol = 1e-04,
  intercept = FALSE
)
}
\arguments{
\item{p}{number of lagged differences}

\item{r}{cointegration rank}

\item{alpha}{initial value for adjustment coefficients}

\item{beta}{initial value for cointegrating vector}

\item{max.iter}{maximum number of iterations}

\item{conv}{convergence parameter}

\item{lambda_gamma}{tuning paramter short-run effects}

\item{rho_omega}{tuning parameter inverse error covariance matrix}

\item{lambda_beta}{tuning paramter cointegrating vector}

\item{cutoff}{cutoff value time series cross-validation approach}

\item{tol}{tolerance parameter glmnet function}

\item{Y}{Response Time Series}

\item{X}{Time Series in Differences}

\item{Z}{Time Series in Levels}
}
\value{
A list containing:
beta: estimate of cointegrating vectors
alpha: estimate of adjustment coefficients
gamma: estimate of short-run effects
omega: estimate of inverse covariance matrix
}
\description{
Main function to perform sparse cointegration
}
