% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsecoint.R
\name{summary.sparsecoint}
\alias{summary.sparsecoint}
\title{Summary function for sparsecoint models}
\usage{
\method{summary}{sparsecoint}(x)
}
\arguments{
\item{x}{The sparsecoint model object}
}
\value{
A summary of the model
}
\description{
Summary function for sparsecoint models
}
